Added section to README that details the acronyms
At the bottom of the README a list of references and the appropriate link.
Then within the document you wrap the acronym in [] to use the link
provided.

Therefore it only has to be changed in one location. This is the current
STRICT markdown syntax that should be used to be XHTML compliant.

The weather is dark and rainy and my coffee mug is empty...